+++ Engine/scenes/level/lvl_npc.h
@@ -28,7 +28,7 @@ public:
 
     //float posX();
     //float posY();
-    void render(float camX, float camY);
+    void render(double camX, double camY);
 };
 
 #endif // LVL_NPC_H
